Refactor with partitionReport.
