+++ assets/js/components/adminbar/AdminBarClicks.js
@@ -29,7 +29,7 @@ import { CORE_USER } from '../../googlesitekit/datastore/user/constants';
 import { CORE_SITE } from '../../googlesitekit/datastore/site/constants';
 import { MODULES_SEARCH_CONSOLE, DATE_RANGE_OFFSET } from '../../modules/search-console/datastore/constants';
 import { calculateChange } from '../../util';
-import { isZeroReport } from '../../modules/search-console/util/is-zero-report';
+import { isZeroReport, partitionReport } from '../../modules/search-console/util';
 import PreviewBlock from '../PreviewBlock';
 import DataBlock from '../DataBlock';
 import sumObjectListValue from '../../util/sum-object-list-value';
@@ -41,6 +41,7 @@ function AdminBarClicks( { WidgetReportZero, WidgetReportError } ) {
 		compare: true,
 		offsetDays: DATE_RANGE_OFFSET,
 	} ) );
+	const dateRangeLength = useSelect( ( select ) => select( CORE_USER ).getDateRangeNumberOfDays() );
 	const reportArgs = {
 		startDate: compareStartDate,
 		endDate,
@@ -64,13 +65,9 @@ function AdminBarClicks( { WidgetReportZero, WidgetReportError } ) {
 		return <WidgetReportZero moduleSlug="search-console" />;
 	}
 
-	// Split the data in two chunks.
-	const half = Math.floor( searchConsoleData.length / 2 );
-	const latestData = searchConsoleData.slice( half );
-	const olderData = searchConsoleData.slice( 0, half );
-
-	const totalClicks = sumObjectListValue( latestData, 'clicks' );
-	const totalOlderClicks = sumObjectListValue( olderData, 'clicks' );
+	const { compareRange, currentRange } = partitionReport( searchConsoleData, { rangeLength: dateRangeLength } );
+	const totalClicks = sumObjectListValue( currentRange, 'clicks' );
+	const totalOlderClicks = sumObjectListValue( compareRange, 'clicks' );
 	const totalClicksChange = calculateChange( totalOlderClicks, totalClicks );
 
 	return (
